Remove tags and Update Get Route
I removed the tags from the database. Curretly finding a solution to how they can be implemented as arrays.
However for the frontend I did not use tags, so I do not need them for now. I will come back and try to find
the solution.
Also the GET route is now functional. Have one test with it as well.